Fixed tool selection and (select tools command as well) and line-ending tool

+++ Svg.Editor.Sample.Forms/Svg.Editor.Sample.Forms/Services/SvgCachingService.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.IO;
 using Svg.Editor.Interfaces;
 using Svg.Interfaces;
 
@@ -8,39 +9,70 @@ namespace Svg.Editor.Sample.Forms.Services
     {
         private readonly Func<string, ISvgSource> _sourceProvider;
 
+        private Lazy<IFileSystem> _fs = new Lazy<IFileSystem>(() => Engine.Resolve<IFileSystem>());
+
         public SvgCachingService()
         {
             _sourceProvider = path => Engine.Resolve<ISvgSourceFactory>().Create(path);
         }
 
-        public void SaveAsPng(string svgFilePath, string nameModifier, Action<SvgDocument> preprocessAction)
+        public string GetCachedPng(string svgFilePath, SaveAsPngOptions options)
         {
+            if (svgFilePath == null) throw new ArgumentNullException(nameof(svgFilePath));
+            if (options == null) throw new ArgumentNullException(nameof(options));
+
             // load svg from FS
             var document = SvgDocument.Open<SvgDocument>(_sourceProvider(svgFilePath));
 
             // apply changes to svg
-            preprocessAction?.Invoke(document);
+            options.PreprocessAction?.Invoke(document);
 
+            var dimension = GetDimension(options);
+            
             // save svg as png
-            using (var bmp = document.DrawAllContents(Engine.Factory.Colors.Transparent))
+            using (var bmp = Engine.Factory.CreateBitmap((int)dimension.Width, (int)dimension.Height))
             {
-                var fs = Engine.Resolve<IFileSystem>();
-                var path = GetCachedPngPath(svgFilePath, nameModifier, fs);
+                document.DrawAllContents(bmp, options.BackgroundColor);
+                var fs = _fs.Value;
+                var path = GetCachedPngPath(svgFilePath, options);
                 if (fs.FileExists(path))
-                    fs.DeleteFile(path);
+                {
+                    // if re-creation is forced
+                    if (options.Force)
+                        fs.DeleteFile(path);
+                    else
+                        return path;
+                }
 
                 using (var stream = fs.OpenWrite(path))
                 {
                     bmp.SavePng(stream);
                 }
+
+                return path;
             }
         }
 
-        public string GetCachedPngPath(string svgFilePath, string nameModifier, IFileSystem fs)
+        public void Clear()
         {
-            var fileName = System.IO.Path.GetFileNameWithoutExtension(svgFilePath);
+            var fs = _fs.Value;
+            var dirPath = fs.PathCombine(fs.GetDefaultStoragePath(), "SvgCache");
+            if (fs.FolderExists(dirPath))
+                fs.DeleteFolder(dirPath);
+        }
+        
+        private string GetCachedPngPath(string svgFilePath, SaveAsPngOptions options)
+        {
+            var fs = _fs.Value;
+            var dim = GetDimension(options);
             fs.EnsureDirectoryExists(fs.PathCombine(fs.GetDefaultStoragePath(), "SvgCache"));
-            return fs.PathCombine(fs.GetDefaultStoragePath(), "SvgCache", string.IsNullOrWhiteSpace(nameModifier) ? $"{fileName}.png" : $"{fileName}_{nameModifier}.png");
+            var filename = options.NamingConvention(svgFilePath, options) ?? $"{Path.GetFileNameWithoutExtension(svgFilePath)}_{(int)dim.Width}px_{(int)dim.Height}px.png";
+            return fs.PathCombine(fs.GetDefaultStoragePath(), "SvgCache", filename);
+        }
+
+        private static SizeF GetDimension(SaveAsPngOptions options)
+        {
+            return options.ImageDimension ?? SizeF.Create(120, 120);
         }
     }
 }
